* Rename cons/filter.c to grep.c   - Some random changes.. with no success   - Grepping console buffer is not working * Cleanup the 'ds' and 'dp' commands   - Implement 'dpa' to attach   - Allow to list childrens of pid   - Also list all attachable pids * Fix segfault by leaking.. needs some cleanup * Fix r_debug_reg_list (NULL) issue
--HG--
rename : libr/cons/filter.c => libr/cons/grep.c

+++ libr/core/config.c
@@ -82,7 +82,7 @@ static int config_asm_bits_callback(void *user, void *data) {
 		}
 	}
 	if (!r_anal_set_bits (&core->anal, node->i_value))
-		eprintf ("asm.arch: Cannot setup '%i' bits analysis engine\n", node->i_value);
+		eprintf ("asm.arch: Cannot setup '%i' bits analysis engine\n", (int)node->i_value);
 	// TODO: change debugger backend bit profile here
 	return ret;
 }
